Speed factor of 14 is too fast. Change to 20.
A speed factor of 14 would mean the 99.99% of the coins would be mined within the first year according to the calculator on https://cryptonotestarter.org/inner.html. A speed factor of 20 gives a more reasonable distribution over time and more closely matches Bitcoin. We will have to honor Valentine's Day another way.